fix some issues for use in module
